Fixed threading data type for 64 bit support
Fixed threading data type for 64 bit support and turned on PThreads in the OSX build config

+++ src/Library/Utilities/Threads/Threads.h
@@ -22,7 +22,7 @@
 
 namespace RISE
 {
-	typedef unsigned int RISETHREADID;
+	typedef unsigned long long RISETHREADID;
 	typedef void* RISEMUTEX;
 	typedef void* RISESLEEP;
 	typedef void* RISESEMAPHORE;
